vconfig:  From old 2006 cvs checkout.
Our web server lost the original, but I found this on an older
server, with file dates from 2006.  I guess this is likely good
enough, but I have not verified it actually works.

The 'ip' tool supports vlan config over netlink, that is preferred
API for last decade or so.

Signed-off-by: Ben Greear <greearb@candelatech.com>